Found bug in IE10 with pushState
